+++ components/scheduler/src/main/java/org/cardanofoundation/ledgersync/scheduler/jobs/OffChainDataScheduler.java
@@ -4,7 +4,7 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
-import org.cardanofoundation.ledgersync.scheduler.service.OffChainDataStoringService;
+import org.cardanofoundation.ledgersync.scheduler.service.OffChainPersistService;
 import org.cardanofoundation.ledgersync.scheduler.OffChainDataProperties;
 import org.cardanofoundation.ledgersync.scheduler.service.OffChainRetryDataErrorService;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -15,8 +15,8 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class OffChainDataScheduler {
 
-    final OffChainDataStoringService offChainDataStoringService;
-    final OffChainRetryDataErrorService offChainDataFetchingErrorService;
+    final OffChainPersistService offChainPersistService;
+    final OffChainRetryDataErrorService offChainRetryDataErrorService;
     final OffChainDataProperties offChainDataProperties;
     final int a = 1;
 
@@ -28,7 +28,7 @@ public class OffChainDataScheduler {
     public void fetchOffChain() {
         log.info("-----------Start job fetch pool offline data-----------");
         final var startTime = System.currentTimeMillis();
-        offChainDataStoringService.validateAndPersistData();
+        offChainPersistService.validateAndPersistData();
         log.info(
                 "----------End job fetch pool offline data, time taken: {} ms----------",
                 (System.currentTimeMillis() - startTime));
@@ -41,7 +41,7 @@ public class OffChainDataScheduler {
         log.info("-----------Start job retry offchain data-----------");
         final var startTime = System.currentTimeMillis();
 
-        offChainDataFetchingErrorService.retryOffChainErrorData();
+        offChainRetryDataErrorService.retryOffChainErrorData();
 
         log.info(
             "----------End job retry offchain data, time taken: {} ms----------",
