--- components/scheduler/src/main/java/org/cardanofoundation/ledgersync/scheduler/service/OffChainDataStoringService.java
@@ -1,10 +1,0 @@
-package org.cardanofoundation.ledgersync.scheduler.service;
-
-public interface OffChainDataStoringService {
-
-    /**
-     * Validate and save the off-chain data
-     */
-    void validateAndPersistData();
-
-}
