+++ components/scheduler/src/main/java/org/cardanofoundation/ledgersync/scheduler/service/OffChainPersistService.java
@@ -1,0 +1,7 @@
+package org.cardanofoundation.ledgersync.scheduler.service;
+
+public interface OffChainPersistService {
+
+    void validateAndPersistData();
+
+}
