+++ components/scheduler/src/main/java/org/cardanofoundation/ledgersync/scheduler/service/offchain/ProcessService.java
@@ -1,0 +1,8 @@
+package org.cardanofoundation.ledgersync.scheduler.service.offchain;
+
+import org.cardanofoundation.ledgersync.consumercommon.entity.OffChainDataCheckpoint;
+
+public interface ProcessService {
+    void process(long startTime, OffChainDataCheckpoint cpType);
+
+}
